requirements.txt
fastapi
uvicorn
sentence-transformers
pillow
requests
torch
